Split componoents out of Details.tsx and added prop typing to each component
